Adds country scores to each relevant country
Each country score is created from the 5 locales with the largest numbers of users per each country. Each locale's localization percentage complete is multiplied by the number of users who use that locale in the country. Then these are are summed for all locales, and finally, divided by the total number of users in a country.